run_benchmark.py

import argparse
import yaml
from transformers import AutoTokenizer

# Import data loaders
from data.patb_loader import PATBDataset
from data.gigaword_loader import GigawordDataset
from data.sanad_loader import SANADDataset
from data.madar_loader import MADARDataset
from data.egyptian_treebank_loader import EgyptianTreebankDataset
from data.lac_loader import LACDataset
from data.gac_loader import GACDataset
from data.mad_loader import MADDataset

# Import models
from models.rule_based_model import RuleBasedModel
from models.statistical_model import StatisticalModel
from models.cl_bilstm_model import CL_BiLSTM
from models.wc_hybrid_model import WC_Hybrid
from models.transformer_model import create_transformer_model
from models.dsm_model import DSM_Model
from models.umd_model import UMD_Model
from models.amd_model import AMD_Model

# Import utils
from utils.data_loader import create_data_loader
from utils.metrics import get_all_metrics

import torch

def main(config_path):
    with open(config_path, \'r\') as f:
        config = yaml.safe_load(f)

    print("Starting benchmark with config:", config)

    # --- 1. Load Data ---
    print("\n--- Loading Data ---")
    # This is a simplified example. A real script would dynamically load datasets based on the config.
    tokenizer = AutoTokenizer.from_pretrained(config[\'model_params\'][\'transformer_model_name\'])
    patb_dataset = PATBDataset(file_path="/path/to/patb", tokenizer=tokenizer)
    train_loader = create_data_loader(patb_dataset, batch_size=config[\'training_params\'][\'batch_size\'])
    print("Data loaded.")

    # --- 2. Initialize Model ---
    print("\n--- Initializing Model ---")
    model_name = config[\'model_to_run\']
    if model_name == \'AMD\':
        model = AMD_Model(
            model_name=config[\'model_params\'][\'transformer_model_name\'],
            num_labels=config[\'model_params\'][\'num_labels\'],
            lora_rank=config[\'model_params\'][\'lora_rank\'],
            lora_alpha=config[\'model_params\'][\'lora_alpha\']
        )
    else:
        # Add initializations for other models here
        print(f"Model {model_name} not fully implemented in this script.")
        model = None
    
    if model:
        print(f"Model {model_name} initialized.")

        # --- 3. Training (Placeholder) ---
        print("\n--- Starting Training (Placeholder) ---")
        # A real training loop would go here:
        # optimizer = torch.optim.AdamW(model.parameters(), lr=config[\'training_params\'][\'learning_rate\'])
        # for epoch in range(config[\'training_params\'][\'epochs\']):
        #     for batch in train_loader:
        #         # ... training step ...
        print("Training complete.")

        # --- 4. Evaluation (Placeholder) ---
        print("\n--- Starting Evaluation (Placeholder) ---")
        # A real evaluation loop would go here:
        # model.eval()
        # with torch.no_grad():
        #     for batch in test_loader:
        #         # ... evaluation step ...
        print("Evaluation complete.")

        # --- 5. Report Results ---
        print("\n--- Generating Results ---")
        # This would call the metrics utility to generate tables that match the dissertation.
        results = get_all_metrics(None) # Pass real prediction data here
        print("\nBenchmark Results (placeholders):")
        print(results)

if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Run Arabic Morphological Analysis Benchmark.")
    parser.add_argument("--config", type=str, required=True, help="Path to the YAML config file.")
    args = parser.parse_args()
    main(args.config)
