requirements.txt
torch==1.9.0
tensorflow==2.6.0
transformers==4.11.3
spacy==3.1.3
peft>=0.3.0
scikit-learn>=0.24.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
ray[tune]>=1.6.0
onnxruntime>=1.9.0
camel-tools>=1.0.0
